pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "google_adk_mongo_session_service"
version = "0.0.1"
description = "Implementation of `google.adk.sessions.base_session_service.BaseSessionService` for storing session information in MongoDB."
readme = "README.md"
requires-python = ">=3.13"
license = { file = "LICENSE" }
authors = [{ name = "Kirill Lutcenko" }]

dependencies = [
    "google-adk>=1.23.0",
    "pymongo>=4.16.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/google_adk_mongo_session_service"]

[tool.uv]

[dependency-groups]
example = [
    "anthropic>=0.43.0",
    "python-dotenv>=1.2.1",
]
